fix(natives): restore S1 observation and cleanup contracts
Restore promised task outcomes, watch cursor and subscription authority, owner-configured wait defaults, and numeric budgets without u32 narrowing. Acknowledge cancelled cleanup independently of pending results and bridge external owner closure to settled resources. Preserve disposal replay, add checked regressions, regenerate bindings, and migrate SDK documentation and demo. Keep the unresolved bounded-total-history versus indefinite exact-replay contract conflict explicit in issues.md rather than inventing retention refusals.

Related #2884

Assistant-model GPT-6-Astra

+++ crates/atomic-natives/src/task_supervisor.rs
@@ -259,12 +259,13 @@ impl NapiTaskSupervisor {
 		})())
 	}
 	/// Registers the observation before returning. Await observeTaskWait separately.
+	/// Accepts the host-resolved numeric budget without u32 narrowing; omission arms no timer.
 	#[napi(ts_return_type = "{ok:true,value:WaitLease}|{ok:false,error:TaskFailure}")]
 	pub fn wait_for_task(
 		&self,
 		env: Env,
 		task: &TaskLease,
-		budget_ms: Option<u32>,
+		budget_ms: Option<f64>,
 		designation: Option<&HostSession>,
 	) -> DoorValue<WaitLease> {
 		DoorValue(
@@ -280,7 +281,7 @@ impl NapiTaskSupervisor {
 		env: Env,
 		task: &TaskLease,
 		host: &HostSession,
-		budget_ms: Option<u32>,
+		budget_ms: Option<f64>,
 	) -> DoorValue<WaitLease> {
 		DoorValue(
 			self
